fix: fixed theme generation overwritting user configs.
Now when a theme file is generated, it's generated betwheen two lines of
text. The "$pywal_gen_text" and the "$pywal_eof_text".

These lines serve to delimit the part of the configuration file that was
generated by pywal_postrun. And can be used inside the script with a sed
command to mark the lines that can actually be overwritten by the
script.

This allows the user to keep their own configuration intact while
pywal_postrun can do the work on theming the right files.