Added extra colorscheme setting functionality.
Lustbg now works with either 0, 1 or 2 arguments.

- if no arguments were passed, lustbg will set the colorscheme with the
dependency script `flust`.

- if one argument was passed, it will check if the argument is either a
path to an image file or if it's a path to a wallust colorscheme
template, and use the apply_colorscheme() or apply_wallpaper() functions
accordingly.

- if two arguments were passed, with the first argument being the path
to an image and the second argument being the path to a wallust
colorscheme template, it will set the image as an wallpaper using waybg
and set the wallust colorscheme that was given by the second argument.