fix(hypr.conf): Fixed terminals being opened with font issues.
Before this commit, opening the terminal emulator (in this case wezterm)
with an extra program to be ran inside would cause certain font
rendering issues on the child process, such as icons or the "ç"
character not being displayed properly. That was because the child
process had limitations because they were not being ran inside a login
shell, which has been fixed.