refactor(nvim/keymaps): Improved `pum_map()`
The `pum_map()` function now requires a table as the first argument.
This allows for finer control over which keymappings will be used for
when the completion popup menu is opened or not.

The table in question requires two entries, with a third optional one.
The first one is the `normal` entry, and it dictates which key(s) will
be pressed when the popup menu is not opened. Likewise the second `pum`
entry dictates which key(s) will be pressed when the popup menu is
opened.
The third `key` entry is optional, it's only purpose is to tell
vim.keymap.set which key should be used to activate the pum mapping. If
that entry is not specified the key(s) set in the `normal` entry will be
used instead.