feat(git/commit-msg): simplify type checking
The checking of commit messages has been simplified.

First a new function was added, the `check()` function. Which really
cleans up the process of grepping commit subjects.
The function works by performing the check
`if echo "$COMMIT_SUBJECT" | grep -q "$1"` under the hood, where
`$1` is just the regex that will be evaluated for the commit linting.

Another change was also made for the "$VALID_COMMIT_TYPES" variable, as
it now also stores the required grep regex to check for it, namely:
`^\(${VALID_COMMIT_TYPES\)`