feat(nvim/lualine): better location module
This commit improves the `location` module by replacing it with a
function that explicitly tells which line/column the cursor is currently
at by adding text together with the numbers.

The default behaviour of the "location" module is to just tell which
line/column the cursor is currently located with numbers separated by a
double colon (e.g.: 100:97). This is pretty confusing as there's no
indication of which number is what.

The new location function module fixes that by specifying which number
is what with a set of strings (e.g.: ln:100 cl:97). This makes the
location of the cursor much more clear within the status bar
information.