feat(nvim/keymaps): better netrw splitting maps
This commit improves the default behaviour of opening files on new
window splits when inside netrw. It does that by defining a new
`netrw_split_file()` global function. That's responsible to detect the
full filepath for the file under the cursor within netrw and using that
to open a new window split.

This commit also adds four new netrw keymaps, those are:
"-": Responsible to open a new vertical split.

"_": Responsible to open a new horizontal split.

"o": This is netrw's builtin keymapping for opening a new horizontal
split. This keymap has been modified to use the new `netrw_split_file()`
function.

"v": This is netrw's builtin keymapping for opening a new vertical
split. This keymap has been modified to use the new `netrw_split_file()`
function.