feat(ftplugin/netrw): add netrw filetype options
This commit adds sensible options to use within netrw. But since this
options are only meant to be used inside netrw and nothing more, they
have been added as a netrw filetype plugin. This ensures that those
options will only be applied when inside a netrw window, since that
window will have the `netrw` filetype.