feat(nvim/mini.surround): <Leader>s as prefix
This commit changes the keymappings for `mini.surround` by making
"<Leader>s" the prefix for each keymap action. This avoids conficts
with the builtin "s" keymapping from Neovim and also allows for better
integration with plugins like `which-key` or `mini.clue` since
`<Leader>s` now is it's own group of keymappings.