feat(nvim/mini.splijoin): <Leader>j as prefix
This commit changes the keymappings for `mini.splitjoin` by making
"<Leader>j" as the prefix for each keymap action. This change allows to
easily add keymaps for the split/join actions without having to worry
about possible conficts with other keymappings and also allows for
better integration with plugins like `which-key` or `mini.clue` since
`<Leader>j` now is it's own group of keymappings.